+++ test/main.tf
@@ -3,20 +3,27 @@ module "thisUbuntu" {
   providers = {
     vsphere = vsphere.homeLab
   }
-  vmCPUs                = 1
-  vmRAM                 = 1024
-  vmDiskSize            = 16
-  vsphereName           = "ubuntu-2204"
-  vmHostName            = "ubuntu-2204"
-  vmHostDomain          = var.vmHostDomain
-  vmRole                = "Test"
+  vmCPUs       = 4
+  vmRAM        = 16384
+  vmDiskSize   = 16
+  vsphereName  = "ubuntu-2204"
+  vmHostName   = "ubuntu-2204"
+  vmHostDomain = var.vmHostDomain
+  vmRole       = "Test"
+  vmAdditionalDisks = {
+    extraDisk = {
+      size          = 64
+      thinProvision = true
+      unitNumber    = 1
+    }
+  }
   vsphereDatacenterName = var.vsphereDatacenterName
   vsphereClusterName    = var.vsphereClusterName
   vsphereNetworkName    = var.vsphereNetworkName
   vsphereFolderName     = "test-vms"
   vsphereDatastoreName  = var.vsphereDatastoreName
   vsphereContentLibrary = null # var.vsphereContentLibrary
-  vsphereTemplateName   = "linux-ubuntu-22.04-lts-main"
+  vsphereTemplateName   = "ubuntu-22.04-latest"
   vsphereResourceName   = null
 }
 
@@ -37,4 +44,9 @@ output "thisUbuntuDatastore" {
 output "thisUbuntuFolder" {
   sensitive = false
   value     = module.thisUbuntu.thisFolder
+}
+
+output "thisTemplate" {
+  sensitive = false
+  value     = module.thisUbuntu.thisTemplate
 }